Move file upload to AddSectionDialog. Remove unneeded deps. Resolve #6627.
