+++ packages/apps/plugins/plugin-stack/package.json
@@ -35,8 +35,6 @@
     "@dxos/util": "workspace:*",
     "@preact/signals-core": "^1.6.0",
     "node-fetch": "^2.6.0",
-    "react-drag-drop-files": "^2.3.8",
-    "react-dropzone": "^14.2.3",
     "react-is": "^18.2.0"
   },
   "devDependencies": {
