+++ packages/apps/plugins/plugin-stack/src/components/AddSectionDialog.tsx
@@ -1,10 +1,11 @@
 //
 // Copyright 2024 DXOS.org
 //
-import React, { useCallback, useState } from 'react';
+import { FilePlus } from '@phosphor-icons/react';
+import React, { useCallback, useRef, useState } from 'react';
 
-import { FolderType, type StackType, SectionType } from '@braneframe/types';
-import { usePlugin, useIntent, LayoutAction } from '@dxos/app-framework';
+import { FolderType, type StackType, SectionType, FileType } from '@braneframe/types';
+import { usePlugin, useIntent, LayoutAction, useResolvePlugin, parseFileManagerPlugin } from '@dxos/app-framework';
 import { create } from '@dxos/echo-schema';
 import { Filter, getSpace, useQuery } from '@dxos/react-client/echo';
 import { Dialog, toLocalizedString, useTranslation } from '@dxos/react-ui';
@@ -33,6 +34,8 @@ export const dataHasAddSectionDialogProps = (data: any): data is { subject: AddS
 export const AddSectionDialog = ({ path, position, stack }: AddSectionDialogProps) => {
   const { t } = useTranslation(STACK_PLUGIN);
   const stackPlugin = usePlugin<StackPluginProvides>(STACK_PLUGIN);
+  const fileManagerPlugin = useResolvePlugin(parseFileManagerPlugin);
+  const fileRef = useRef<HTMLInputElement | null>(null);
   const { dispatch } = useIntent();
   const space = getSpace(stack);
   const [folder] = useQuery(space, Filter.schema(FolderType));
@@ -55,6 +58,18 @@ export const AddSectionDialog = ({ path, position, stack }: AddSectionDialogProp
     },
     [stack, stack.sections, path, position, dispatch],
   );
+
+  const handleFileUpload =
+    fileManagerPlugin?.provides.file.upload && space
+      ? async (file: File) => {
+          const filename = file.name.split('.')[0];
+          const info = await fileManagerPlugin.provides.file.upload?.(file, space);
+          if (info) {
+            const obj = create(FileType, { type: file.type, title: filename, filename, cid: info.cid });
+            handleAdd(obj);
+          }
+        }
+      : undefined;
 
   return (
     <Dialog.Content>
@@ -81,8 +96,26 @@ export const AddSectionDialog = ({ path, position, stack }: AddSectionDialogProp
               </SearchList.Item>
             );
           })}
+          {handleFileUpload && (
+            <SearchList.Item
+              value={t('upload file label')}
+              classNames='flex items-center gap-2 pli-2'
+              onSelect={() => fileRef.current?.click()}
+            >
+              <FilePlus />
+              <span className='grow truncate'>{t('upload file label')}</span>
+            </SearchList.Item>
+          )}
         </SearchList.Content>
       </SearchList.Root>
+      {handleFileUpload && (
+        <input
+          type='file'
+          className='sr-only'
+          ref={fileRef}
+          onChange={({ target: { files } }) => files && handleFileUpload(files[0])}
+        />
+      )}
     </Dialog.Content>
   );
 };
